Add more recoverWith tests
