+++ cyclops/src/test/java/cyclops/control/future/futureOverwriteIssue/AuthorizationService3.java
@@ -1,4 +1,4 @@
-package cyclops.control.futureOverwriteIssue;
+package cyclops.control.future.futureOverwriteIssue;
 
 
 import cyclops.control.Either;
