+++ cyclops/src/test/java/cyclops/control/future/futureOverwriteIssue/SuccessURLDataFileMetadata.java
@@ -1,15 +1,10 @@
-package cyclops.control.futureOverwriteIssue;
+package cyclops.control.future.futureOverwriteIssue;
 
 import cyclops.control.Either;
 import lombok.Getter;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.net.URL;
-import java.net.URLConnection;
-import java.nio.charset.StandardCharsets;
-import java.util.stream.Collectors;
 
 
 @Getter
