+++ cyclops/src/test/java/cyclops/control/future/futureOverwriteIssue/User.java
@@ -1,4 +1,4 @@
-package cyclops.control.futureOverwriteIssue;
+package cyclops.control.future.futureOverwriteIssue;
 
 import cyclops.data.Seq;
 import lombok.AllArgsConstructor;
