+++ cyclops/src/test/java/cyclops/control/ior/IorBothTest.java
@@ -1,6 +1,7 @@
-package cyclops.control;
+package cyclops.control.ior;
 
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.nullValue;
@@ -13,36 +14,51 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import cyclops.control.Ior;
 import cyclops.control.Option;
-import cyclops.control.Either;
 import org.junit.Before;
 import org.junit.Test;
 
-public class EitherRightTest {
+public class IorBothTest {
 
-	Either<FileNotFoundException,Integer> success;
+	Ior<FileNotFoundException,Integer> success;
 	final Integer value = 10;
-	
-	
-	public Either<FileNotFoundException,String> load(String filename){
-		return Either.right("test-data");
+
+
+	public Ior<FileNotFoundException,String> load(String filename){
+		return Ior.both(new FileNotFoundException(),"test-data");
 	}
-	
+
 	public void process(){
-		
-		Either<FileNotFoundException,String> attempt = load("data");
-		
+
+		Ior<FileNotFoundException,String> attempt = load("data");
+
 		attempt.map(String::toUpperCase)
 				.peek(System.out::println);
-	
+
 	}
-	
+
 	@Before
 	public void setup(){
-		success = Either.right(10);
+		success = Ior.both(new FileNotFoundException(),10);
 	}
+	@Test
+	public void bimap(){
 
-
+	    Ior<RuntimeException,Integer> mapped = success.bimap(e->new RuntimeException(), d->d+1);
+	    assertThat(mapped.orElse(-10),equalTo(11));
+	    assertThat(mapped.swap().orElse(null),instanceOf(RuntimeException.class));
+	}
+	Throwable capT;
+	int capInt=0;
+	@Test
+    public void bipeek(){
+       capT =null;
+       capInt=0;
+         success.bipeek(e->capT=e, d->capInt=d);
+        assertThat(capInt,equalTo(10));
+        assertThat(capT,instanceOf(FileNotFoundException.class));
+    }
 
 	@Test
 	public void testGet() {
@@ -56,24 +72,24 @@ public class EitherRightTest {
 
 	@Test
 	public void testMap() {
-		assertThat(success.map(x->x+1),equalTo(Either.right(value+1)));
+		assertThat(success.map(x->x+1).get(),equalTo(Ior.right(value+1).get()));
 	}
 
 	@Test
 	public void testFlatMap() {
-		assertThat(success.flatMap(x-> Either.right(x+1)),equalTo(Either.right(value+1)));
+		assertThat(success.flatMap(x->Ior.right(x+1)).get(),equalTo(Ior.right(value+1).get()));
 	}
 
 	@Test
 	public void testFilter() {
-		assertThat(success.filter(x->x>5),equalTo(Option.some(value)));
+		assertThat(success.filter(x->x>5),equalTo(Ior.right(value).toOption()));
 	}
 	@Test
 	public void testFilterFail() {
 		assertThat(success.filter(x->x>15),equalTo(Option.none()));
 	}
 
-	
+
 
 	@Test
 	public void testOrElse() {
@@ -82,6 +98,7 @@ public class EitherRightTest {
 
 	@Test
 	public void testOrElseGet() {
+
 		assertThat(success.orElseGet(()->30),equalTo(value));
 	}
 
@@ -98,7 +115,8 @@ public class EitherRightTest {
 
 	@Test
 	public void testIsSuccess() {
-		assertTrue(success.isRight());
+		assertTrue(success.isBoth());
+		assertFalse(success.isRight());
 	}
 
 	@Test
@@ -112,6 +130,6 @@ public class EitherRightTest {
 		assertThat(valueCaptured,is(10));
 	}
 	Exception errorCaptured;
-	
+
 
 }
