+++ cyclops/src/test/java/cyclops/control/ior/IorOrElseValueTest.java
@@ -1,0 +1,39 @@
+package cyclops.control.ior;
+
+import com.oath.cyclops.types.OrElseValue;
+import cyclops.control.AbstractOrElseValueTest;
+import cyclops.control.Ior;
+import cyclops.control.Try;
+
+public class IorOrElseValueTest extends AbstractOrElseValueTest {
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> of(int value) {
+        return (OrElseValue) Ior.right(value);
+    }
+
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty2() {
+        return (OrElseValue) Ior.left(new Exception());
+    }
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty3() {
+        return (OrElseValue) Ior.left(new Exception());
+    }
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty4() {
+        return (OrElseValue)Ior.left(new Exception());
+    }
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty1() {
+        return (OrElseValue)Ior.left(new Exception());
+    }
+
+    @Override
+    public boolean isLazy() {
+        return false;
+    }
+}
