+++ cyclops/src/test/java/cyclops/control/ior/IorPrimaryTest.java
@@ -1,7 +1,6 @@
-package cyclops.control;
+package cyclops.control.ior;
 
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.nullValue;
@@ -14,17 +13,19 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import cyclops.control.Ior;
+import cyclops.control.Option;
 import org.junit.Before;
 import org.junit.Test;
 
-public class IorBothTest {
+public class IorPrimaryTest {
 
 	Ior<FileNotFoundException,Integer> success;
 	final Integer value = 10;
 
 
 	public Ior<FileNotFoundException,String> load(String filename){
-		return Ior.both(new FileNotFoundException(),"test-data");
+		return Ior.right("test-data");
 	}
 
 	public void process(){
@@ -38,24 +39,24 @@ public class IorBothTest {
 
 	@Before
 	public void setup(){
-		success = Ior.both(new FileNotFoundException(),10);
+		success = Ior.right(10);
 	}
 	@Test
-	public void bimap(){
+    public void bimap(){
 
-	    Ior<RuntimeException,Integer> mapped = success.bimap(e->new RuntimeException(), d->d+1);
-	    assertThat(mapped.orElse(-10),equalTo(11));
-	    assertThat(mapped.swap().orElse(null),instanceOf(RuntimeException.class));
-	}
-	Throwable capT;
-	int capInt=0;
-	@Test
+        Ior<RuntimeException,Integer> mapped = success.bimap(e->new RuntimeException(), d->d+1);
+        assertThat(mapped.get(),equalTo(Option.some(11)));
+        assertTrue(mapped.isRight());
+    }
+    Throwable capT;
+    int capInt=0;
+    @Test
     public void bipeek(){
        capT =null;
        capInt=0;
          success.bipeek(e->capT=e, d->capInt=d);
         assertThat(capInt,equalTo(10));
-        assertThat(capT,instanceOf(FileNotFoundException.class));
+        assertThat(capT,nullValue());
     }
 
 	@Test
@@ -70,17 +71,17 @@ public class IorBothTest {
 
 	@Test
 	public void testMap() {
-		assertThat(success.map(x->x+1).get(),equalTo(Ior.right(value+1).get()));
+		assertThat(success.map(x->x+1),equalTo(Ior.right(value+1)));
 	}
 
 	@Test
 	public void testFlatMap() {
-		assertThat(success.flatMap(x->Ior.right(x+1)).get(),equalTo(Ior.right(value+1).get()));
+		assertThat(success.flatMap(x->Ior.right(x+1)),equalTo(Ior.right(value+1)));
 	}
 
 	@Test
 	public void testFilter() {
-		assertThat(success.filter(x->x>5),equalTo(Ior.right(value).toOption()));
+		assertThat(success.filter(x->x>5),equalTo(Option.some(value)));
 	}
 	@Test
 	public void testFilterFail() {
@@ -96,7 +97,6 @@ public class IorBothTest {
 
 	@Test
 	public void testOrElseGet() {
-
 		assertThat(success.orElseGet(()->30),equalTo(value));
 	}
 
@@ -113,8 +113,7 @@ public class IorBothTest {
 
 	@Test
 	public void testIsSuccess() {
-		assertTrue(success.isBoth());
-		assertFalse(success.isRight());
+		assertTrue(success.isRight());
 	}
 
 	@Test
