+++ cyclops/src/test/java/cyclops/control/ior/IorSecondaryTest.java
@@ -1,37 +1,54 @@
-package cyclops.control;
+package cyclops.control.ior;
 
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 import java.io.FileNotFoundException;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import cyclops.control.Ior;
 import cyclops.control.Option;
-import cyclops.control.Either;
 import org.junit.Before;
 import org.junit.Test;
 
-public class EitherLeftTest {
+public class IorSecondaryTest {
 
-	Either<FileNotFoundException,Integer> failure;
+	Ior<FileNotFoundException,Integer> failure;
 	FileNotFoundException error = new FileNotFoundException();
 	@Before
 	public void setup(){
-		failure = Either.left(error);
+		failure = Ior.left(error);
 	}
+	@Test
+    public void bimap(){
+
+        Ior<RuntimeException,Integer> mapped = failure.bimap(e->new RuntimeException(), d->d+1);
+        assertTrue(mapped.isLeft());
+        assertThat(mapped.swap().orElse(null),instanceOf(RuntimeException.class));
+    }
+    Throwable capT;
+    int capInt=0;
+    @Test
+    public void bipeek(){
+       capT =null;
+       capInt=0;
+         failure.bipeek(e->capT=e, d->capInt=d);
+        assertThat(capInt,equalTo(0));
+        assertThat(capT,instanceOf(FileNotFoundException.class));
+    }
 
 
 	@Test
 	public void testOf() {
 		assertNotNull(failure);
 	}
-
-
 
 	@Test
 	public void testMap() {
@@ -40,7 +57,7 @@ public class EitherLeftTest {
 
 	@Test
 	public void testFlatMap() {
-		assertThat(failure.flatMap(x-> Either.right(10)),equalTo(failure));
+		assertThat(failure.flatMap(x->Ior.right(10)),equalTo(failure));
 	}
 
 	@Test
@@ -48,9 +65,7 @@ public class EitherLeftTest {
 		assertThat(failure.filter(x->x==10),equalTo(Option.none()));
 	}
 
-	
 
-	
 	@Test
 	public void testOrElse() {
 		assertThat(failure.orElse(10),equalTo(10));
@@ -85,12 +100,12 @@ public class EitherLeftTest {
 	Integer value = null;
 	@Test
 	public void testForeach() {
-		
+
 		failure.forEach(v -> value = v);
 		assertThat(value,is(nullValue()));
 	}
 
-	
+
 	Object errorCaptured;
 	@Test
 	public void testForeachFailed() {
@@ -99,6 +114,6 @@ public class EitherLeftTest {
 		assertThat(error,equalTo(errorCaptured));
 	}
 
-	
+
 
 }
