+++ cyclops/src/test/java/cyclops/control/trytests/TryOrElseValueTest.java
@@ -1,0 +1,40 @@
+package cyclops.control.trytests;
+
+import com.oath.cyclops.types.OrElseValue;
+import cyclops.control.AbstractOrElseValueTest;
+import cyclops.control.Try;
+import cyclops.control.Validated;
+import cyclops.data.NonEmptyList;
+
+public class TryOrElseValueTest extends AbstractOrElseValueTest {
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> of(int value) {
+        return (OrElseValue) Try.success(value);
+    }
+
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty2() {
+        return (OrElseValue) Try.failure(new Exception());
+    }
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty3() {
+        return (OrElseValue) Try.failure(new Exception());
+    }
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty4() {
+        return (OrElseValue)Try.failure(new Exception());
+    }
+
+    @Override
+    public OrElseValue<Integer, OrElseValue<Integer, ?>> empty1() {
+        return (OrElseValue)Try.failure(new Exception());
+    }
+
+    @Override
+    public boolean isLazy() {
+        return false;
+    }
+}
